+++ backend/app/schemas/intelligence.py
@@ -7,8 +7,8 @@ from .common import APIModel
 class IntelligenceEventIn(APIModel):
     route: str | None = None
     event_type: str
-    anomaly_score: Decimal | None = Field(default=None, ge=0)
-    pressure_score: Decimal | None = Field(default=None, ge=0)
+    anomaly_score: float | None = Field(default=None, ge=0)
+    pressure_score: float | None = Field(default=None, ge=0)
     shock_status: str | None = None
     explanation: str | None = None
     affected_sources: list[str] = Field(default_factory=list)
